requirements/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==24.2.0
Brotli==1.1.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cryptography==43.0.0
cssselect2==0.7.0
cycler==0.12.1
Django==5.0.8
django-jet-reboot==1.3.7
django-model-utils==4.5.1
djangorestframework==3.15.2
fonttools==4.53.1
frozenlist==1.4.1
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.2
html5lib==1.1
httplib2==0.22.0
idna==3.7
kiwisolver==1.4.5
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
multidict==6.0.5
numpy==2.0.1
oscrypto==1.3.0
packaging==24.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.27.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydyf==0.11.0
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.9.0
pyparsing==3.1.2
pypdf==4.3.1
pyphen==0.16.0
pypng==0.20220715.0
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
qrcode==7.4.2
reportlab==4.2.2
requests==2.32.3
rsa==4.9
six==1.16.0
sqlparse==0.5.1
svglib==1.5.1
tinycss2==1.3.0
twilio==9.2.4
typing_extensions==4.12.2
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.2.2
weasyprint==62.3
webencodings==0.5.1
xhtml2pdf==0.2.16
yarl==1.9.7
zopfli==0.2.3
gunicorn